requirements.txt
aiogram
ydb
python-dotenv